lib: split os_system_run
This commit splits os_system_run into two distinct functions: os_system_run
and os_system_run_wait. The first runs the command in the background and the
second waits for the command to finish. Output buffers can be null for both
functions. The background function uses file handles instead of char buffers
so that command output can be piped to a file and read later.