build: improve header generation
This commit improves the header generation process by moving it out of the
source directory and into the binary directory, as is expected for builds. In
addition, the header files are no longer found under /build-sys, instead
they're now in /header. Instead of using bash/batch scripts, the header is
created using a CMake script that runs before building the library itself.
Finally, the wrapped build uses this same header generation method so as to be
more consistent and easier to follow.