header: improve code re-use
This commit changes how the header is generated for wrapped builds so as to
re-use more of the same code as the non-wrapped builds. In the new system,
only wrapped functions go in os_wrap.h.in as opposed to wrapped functions and
the contents of os_bot.h.in.