libuuid: pull in a BSD implementation of uuid
The system libuuid is GPL and uuid is required by the osal.  The osal
must support static compilation which is incompatible with GPL
licensing.  Pull in another BSD implmenation suitable for static
compilation.  This commit pulls in version 1.0.3 from:
	https://sourceforge.net/projects/libuuid/
A few small fixes were made to the source to fix compilation warnings.

Signed-off-by: Paul Barrette <paul.barrette@windriver.com>