posix: remove arguments to macro for printf
Using a macro with printf definitions, requires that a caller includes
at least 1 parameter (i.e. the statement of printf( "failure" ); would
not be allowed) do to a limitation of the _VA_ARGS usage and C99.  This
patch updates the macro to not check the parameters in order to remove
this limitation.

Signed-off-by: Keith Holman <keith.holman@windriver.com>