posix: revamp code base to be easier to maintain
This fix updates the code base to be easier to maintain.  Previously,
the code base included "*_top.h.in", "*_bot.h.in", "*_wrappers.h",
"*_macros.h" files.  How all these files interacted was confusing.  This
patch uses cmake to use 1 file and parse preprocessor defintions for
this purpose making it easier to maintain.  Additionally, this change
updates code documentation and installation routines to be better
supported.  This change is for linux (with posix) systems and Windows.
Future patches, will be included for other operating systems.

Signed-off-by: Keith Holman <keith.holman@windriver.com>