build: add include directory location for uuid.h
"uuid/uuid.h" was not being found in the repository, instead it was
looking for it in "/usr/include", which is undesired on posix systems.
This patch updates the build to the current directory and an include
search path.  Additionally, this patch also adds a cmake generated file
to the .gitignore file.

Signed-off-by: Keith Holman <keith.holman@windriver.com>