build: fix to compile with CMake 3.11+
fixes: #20

The signature for check_include_files has changed slightly in CMake
3.11 previously the function would handle multiple header files passed
in as a list with the `${ARGN}` variable.  However, now the new
signature requires the list to be an explicit string.  This fix updates
the code to pass multiple header files in as a string list.  This also
works for previous versions of cmake.

Signed-off-by: Keith Holman <keith.holman@windriver.com>