vxworks: add os_vxworks.h into build
This patch uses the older `os_vxworks.h` file into the vxworks build.
Replacing the new `vxworks_os.h` file.  The changes captured in the new
file are put into the older definition file.  The purpose of this change
is to make the VxWorks build consistent with other operating system
specific header files.

Signed-off-by: Keith Holman <keith.holman@windriver.com>