build: output different library names on windows
On Windows both static and shared libraries have the same import
extension ".lib".  If the target has the same output file then one
library is replaced by a second immediately after building.  This patch
only sets them to have the same name for non-windows builds.  On these
platforms static libraries and shared libraries have the different
extensions to differeniate them.  For example, .so for shared libraries
and .a for static libraries on Linux.

Signed-off-by: Keith Holman <keith.holman@windriver.com>